Fixed player visual bug

+++ FallSlime_LITTLE_BUSTERS/Motor2D/j1Render.cpp
@@ -226,8 +226,8 @@ bool j1Render::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section,
 	uint scale = App->win->GetScale();
 
 	SDL_Rect rect;
-	rect.x = (int)((camera.w * (zoom - 1)) / 2) + (-camera.x * speed) + x * scale;
-	rect.y = (int)((camera.h * (zoom - 1)) / 2) + (-camera.y * speed) + y * scale;
+	rect.x = (int)((camera.w * (zoom - 1)) / 2) + (int)(-camera.x * speed) + x * scale;
+	rect.y = (int)((camera.h * (zoom - 1)) / 2) + (int)(-camera.y * speed) + y * scale;
 
 	if (section != NULL)
 	{
